Add text against exact qs and sat unit tests in own file.
