+++ components/scream/src/physics/p3/tests/CMakeLists.txt
@@ -26,6 +26,7 @@ set(P3_TESTS_SRCS
   p3_calc_liq_relaxation_timescale_unit_tests.cpp
   p3_ice_nucleation_unit_tests.cpp
   p3_ice_melting_unit_tests.cpp
+  p3_saturation_unit_tests.cpp
   p3_droplet_activation_unit_tests.cpp
   p3_ice_cldliq_wet_growth_unit_tests.cpp
   p3_incloud_mixingratios_unit_tests.cpp
