+++ components/scream/src/physics/p3/tests/p3_unit_tests.cpp
@@ -23,236 +23,6 @@ namespace unit_test {
  * Unit-tests for p3_functions.
  */
 template <typename D>
-struct UnitWrap::UnitTest<D>::TestP3Saturation
-{
-
-  KOKKOS_FUNCTION  static void saturation_tests(const Scalar& temperature, const Scalar& pressure, const Scalar& correct_sat_ice_p,
-    const Scalar& correct_sat_liq_p, const Scalar&  correct_mix_ice_r, const Scalar& correct_mix_liq_r, int& errors ){
-
-    //Convert Scalar inputs to Spacks because that's what polysvp1 and qv_sat expect as inputs.
-    //--------------------------------------
-    const Spack temps(temperature);
-    const Spack pres(pressure);
-
-    //Get values from polysvp1 and qv_sat to test against "correct" values
-    //--------------------------------------
-    Spack sat_ice_p = Functions::polysvp1(temps, true);
-    Spack sat_liq_p = Functions::polysvp1(temps, false);
-    Spack mix_ice_r = Functions::qv_sat(temps, pres, true);
-    Spack mix_liq_r = Functions::qv_sat(temps, pres, false);
-
-    // The correct results were computed with double precision, so we need
-    // significantly greater tolerance for single precision.
-    //Scalar tol = (util::is_single_precision<Scalar>::value || util::OnGpu<ExeSpace>::value) ? C::Tol*100 : C::Tol;
-    //PMC comment: C::Tol already has logic to make it 2e-5 if single precision, so multiplying by another 100 seems
-    //gratuitous. Also not sure why coarse tol is used if GPU. 
-    const Scalar tol = C::Tol;
-    const Scalar sym_tol_p = 10.;
-    const Scalar sym_tol_q = 1e-3;
-
-    const Scalar RV = C::RV;
-    const Scalar RhoLiq = C::RhoH2O;
-    const Scalar RhoIce = C::RhoIce;
-    const Scalar LatVap = C::LatVap;
-    const Scalar LatIce = C::LatIce;
-
-    printf("===============================\n");
-    printf("Checking T = %f",temperature," and Pres = %e\n",pressure);
-    printf("===============================\n");
-
-    //PMC note: only checking 1st entry in packs below b/c packs were created
-    //by copying a scalar up to pack size.
-      
-    // Test vapor pressure against approximate symbolic implementation
-    // ---------------------------------------------------------
-    Scalar esi_exact=polysvp1_exact(temperature,1);
-    Scalar esl_exact=polysvp1_exact(temperature,0);
-
-    if ( std::abs(sat_ice_p[0] - esi_exact ) > sym_tol_p ) {
-      printf("  Diff btwn analytic esi=%e and computed=%e bigger than sym_tol=%e\n",esi_exact,sat_ice_p[0],sym_tol_p);
-      errors++;}
-    if ( std::abs(sat_liq_p[0] - esl_exact ) > sym_tol_p ) {
-      printf("  Diff btwn analytic esl=%e and computed=%e bigger than sym_tol=%e\n",esl_exact,sat_liq_p[0],sym_tol_p);
-      errors++;}
-
-    // Test vapor pressure against Flatau's impl of Wexler:
-    // ---------------------------------------------------------
-    // First calculate condition number Cond=x*f'(x)/f(x), which gives the growth in roundoff error
-    // expected from the function. Here, x=temperature, f(x) = sat_X_p, and
-    // f'(x) = L/T*sat_X_p/(Rv*T - sat_X_p/rho_{liq or ice}) from Clausius-Clapeyron.
-
-    const Scalar Cond_ice_p=(LatVap+LatIce)*sat_ice_p[0]/(RV*temps[0] - sat_ice_p[0]/RhoIce);
-    const Scalar Cond_liq_p=LatVap*sat_liq_p[0]/(RV*temps[0] - sat_liq_p[0]/RhoLiq);
-      
-    // Now check that computed vs expected values are small enough.     
-    if ( std::abs(sat_ice_p[0] - correct_sat_ice_p ) > Cond_ice_p*tol ) {
-      printf("  esi: abs(calc-expected),cond*tol=%e %e\n",std::abs(sat_ice_p[0] - correct_sat_ice_p ),tol*Cond_ice_p );
-      errors++;}
-    if (std::abs(sat_liq_p[0] - correct_sat_liq_p) > Cond_liq_p*tol)  {
-      printf("  esl: abs(calc-expected),cond*tol=%e %e\n",std::abs(sat_liq_p[0] - correct_sat_liq_p ),tol*Cond_liq_p);
-      errors++;}
-
-    // Test vapor pressure against exact symbolic implementation
-    // ---------------------------------------------------------
-    Scalar qsi_exact=qvsat_exact(temperature,pressure,1);
-    Scalar qsl_exact=qvsat_exact(temperature,pressure,0);
-
-    if ( std::abs(mix_ice_r[0] - qsi_exact ) > sym_tol_q ) {
-      printf("Diff btwn analytic qsi=%e and computed=%e bigger than sym_tol=%e\n",qsi_exact,mix_ice_r[0],sym_tol_q);
-      errors++;}
-    if ( std::abs(mix_liq_r[0] - qsl_exact ) > sym_tol_q ) {
-      printf("Diff btwn analytic qsl=%e and computed=%e bigger than sym_tol=%e\n",qsl_exact,mix_liq_r[0],sym_tol_q);
-      errors++;}
-
-    //Test mixing-ratios against Wexler approx: 
-    // -------------------
-    // Again, compute condition # Cond=x*f'(x)/f(x), with x=temperature, f(x)=mix_X_r[0], and
-    // f'(x) = L*mix_X_r[0]/(Rv*temperature**2.) from Clausius-Clapeyron. Nice cancelation leaves:
-
-    const Scalar Cond_ice_r=(LatVap+LatIce)/(RV*temps[0]);
-    const Scalar Cond_liq_r=LatVap/(RV*temps[0]);
-      
-    // Now check that computed vs expected values are small enough.
-    if (std::abs(mix_ice_r[0] -  correct_mix_ice_r) > Cond_ice_r*tol ) {
-      printf("qsi: abs(calc-expected)=%e %e\n",std::abs(mix_ice_r[0] -  correct_mix_ice_r),tol*Cond_ice_r);
-      errors++;}
-    if (std::abs(mix_liq_r[0] -  correct_mix_liq_r) > Cond_liq_r*tol ) {
-      printf("qsl: abs(calc-expected)=%e %e\n",std::abs(mix_liq_r[0] -  correct_mix_liq_r),tol*Cond_liq_r);
-      errors++;}
-
-  }
-
-  ////////////////////////////////////////PETER HACKED IN STUFF
-  static Scalar polysvp1_exact(const Scalar& t_atm, const bool& ice)
-  {
-    // Compute saturation vapor pressure (es) symbolically.
-    // des/dT = L/[T*(1/rho_vap - 1/rho_liq] by Clausius clapeyron 
-    // (Wallace + Hobbes eq 2.93). 1/rho_liq<<1/rho_vap =>
-    // des/dT \approx L/[T*Rv*T/es] = L*es/(Rv*T**2.)
-    // Taking the integral between some T0 and a particular T yields:
-    // es(T) \approx es(T0)*exp{-L/Rv*(1/T - 1/T0)}.
-
-    const auto tmelt = C::Tmelt;
-    const auto RV = C::RV;
-    const auto LatVap = C::LatVap;
-    const auto LatIce = C::LatIce;
-
-    Scalar es_T0;
-    Scalar L;
-    
-    //Below, I'm taking es_T0 as equal to the constant term in
-    //Flatau et al 1992 table 4. Would be better to use exact obs
-    //value at tmelt, but Flatau is probably close enough (and provides
-    //better match to P3's implementation!
-    if ((t_atm < tmelt) && ice) {
-      es_T0=611.147274;
-      L=LatVap+LatIce;
-    } else{
-      es_T0=611.239921;
-      L=LatVap;
-    }
-    
-    Scalar result = es_T0*std::exp( -L/RV*(1/t_atm - 1/tmelt) );
-
-    //printf("T,is_ice,es=%e %d %e\n",t_atm,ice,result);
-    
-    return result;
-  }
-
-  static Scalar qvsat_exact(const Scalar& t_atm, const Scalar& p_atm, const bool& ice)
-  {
-    // Compute saturation mixing ratio qs symbolically.
-    // dqs/dT = L*qs/(Rv*T**2.) by Clausius Clapeyron.
-    // Taking the integral between some T0 and a particular T yields:
-    // qs(T) = qs(T0)*exp{-L/Rv*(1/T - 1/T0)}.
-    // Unlike for saturation vapor pressure, this formulation *is*
-    // perfectly exact. It isn't used in numerical models just because
-    // it is computationally expensive. 
-
-    const auto tmelt = C::Tmelt;
-    const auto RV = C::RV;
-    const auto LatVap = C::LatVap;
-    const auto LatIce = C::LatIce;
-    const auto ep_2 = C::ep_2;
-
-    Scalar es_T0;
-    Scalar L;
-
-    //Below, I'm taking es_T0 as equal to the constant term in
-    //Flatau et al 1992 table 4 and turning it into qs_T0 using the
-    //exact formula for es->qs. Would be better to use exact obs
-    //value at tmelt, but Flatau is probably close enough (and provides
-    //better match to P3's implementation!
-    if ((t_atm < tmelt) && ice) {
-      es_T0=611.147274; 
-      L=LatVap+LatIce;
-    } else{
-      es_T0=611.239921;
-      L=LatVap;
-    }
-
-    // 
-    Scalar qs_T0 = ep_2 * es_T0 / std::max(p_atm-es_T0, sp(1.e-3));
-    Scalar result = qs_T0*std::exp( -L/RV*(1/t_atm - 1/tmelt) );
-
-    //printf("T,is_ice,qs=%e %d %e\n",t_atm,ice,result);
-
-    return result;
-  }
-
-  ////////////////////////////////////////END HACK
-  static void run()
-  {
-    /*Originally written by Kyle Pressel, updated by Peter Caldwell on 4/5/20. 
-     *This code tests polysvp1 and qv_sat at 0 degrees C, at a very cold T, and at a very hot T
-     *to make sure our impl gets the same answer as Flatau et al 1992:
-     *(https://journals.ametsoc.org/doi/abs/10.1175/1520-0450%281992%29031%3C1507%3APFTSVP%3E2.0.CO%3B2
-     *For 0 degrees C, polysvp values can be read directly from Flatau. For other cases, I independently 
-     *coded up the Flatau scheme in python and used it to derive the expected values. My python code is
-     *in https://github.com/E3SM-Project/scream-docs.git analysis-scripts/test_qv_sat.py
-     */
-
-    
-    int nerr = 0;
-    TeamPolicy policy(util::ExeSpaceUtils<ExeSpace>::get_default_team_policy(1, 1));
-    Kokkos::parallel_reduce("TestTableIce::run", policy, KOKKOS_LAMBDA(const MemberType& team, int& errors) {
-
-      errors = 0;
-      const auto tmelt = C::Tmelt;
-      
-      // Just Freezing Case: Test values @ 273.15K @ 1e5 Pa
-      //---------------------------------------
-      // This is the nicest test b/c polysvp1 is a polynomial fit around (T-273.15)
-      // so T=273.15 collapses back to the intercept coefficient which can be read 
-      // directly from the RHS of table 4 of Flatau et al 1992.
-      // Note that ice values are identical to liquid values b/c C++ uses liq val for T>=0 C.
-
-      saturation_tests(tmelt, 1e5, 611.23992100000009, 611.23992100000009,
-		       0.0038251131382843278, 0.0038251131382843278, errors);
-      
-      // Cold Case: Test values @ 243.15K @ 1e5 Pa
-      //---------------------------------------
-      saturation_tests(243.15, 1e5, 38.024844602056795, 51.032583257624964,
-		       0.00023659331311441935, 0.00031756972127516819, errors);
-
-      //Warm Case: Test values @ 303.15 @ 1e5 Pa
-      //---------------------------------------
-      saturation_tests(303.15, 1e5, 4245.1933273786717, 4245.1933273786717,
-		       0.027574442204332306, 0.027574442204332306, errors);
-
-      //Change Pressure Case: Test values @ 243.15 @ 500 mb
-      //---------------------------------------
-      saturation_tests(243.15, 5e4, 38.024844602056795, 51.032583257624964,
-		       0.00047336669164733106, 0.00063546390177500586, errors);
-      
-    }, nerr);
-
-    Kokkos::fence();
-    REQUIRE(nerr == 0);
-  }
-}; //end of TestP3Saturation struct
-
-template <typename D>
 struct UnitWrap::UnitTest<D>::TestP3Conservation
 {
 
@@ -1557,11 +1327,6 @@ struct UnitWrap::UnitTest<D>::TestP3FunctionsImposeMaxTotalNi
 }//namespace scream
 
 namespace {
-
-TEST_CASE("p3_functions", "[p3_functions]")
-{
-  scream::p3::unit_test::UnitWrap::UnitTest<scream::DefaultDevice>::TestP3Saturation::run();
-}
 
 TEST_CASE("p3_conservation_test", "[p3_conservation_test]"){
   scream::p3::unit_test::UnitWrap::UnitTest<scream::DefaultDevice>::TestP3Conservation::run();
